* Initial import of the th0rpe's fuzzer   - Based on new !loop debugger command and a LUA script   - Coded and tested on windows * Add fuzzer-loop.lua script * New lua command to return integer value instead of string   - regval command for debugger (needs refactoring)

+++ src/dbg/parser.c
@@ -32,7 +32,7 @@
 #include "../utils.h"
 
 extern struct regs_off roff[];
-extern int get_reg(char *reg);
+extern unsigned long get_reg(char *reg);
 
 #define ishexa(c) ((c >='0' && c <= '9') || \
 		  (tolower(c) >='a' && tolower(c) <= 'f'))
